requirements.txt
youtube_transcript_api
translators
deep-translator
google.generativeai
streamlit
pathlib
python-dotenv